Commands run as a subprocess to capture output
Only capturing output when there's a log file. Otherwise you can watch
the commands output as it happens in the terminal.

Not sure how to capture and subprocess output as it happens, would be
better, think you can set a file handle I'll play with that next to
speed up logging.