Launcher
Put together a little launcher file together that lets you set some
command aliases and is automatically executable and pointing at your
local confply install.

I really don't want to put confply in my install dirs or on my path, I
think it'll be a better tool if it's local to repos and linked to a
github or whatever. Less likely to break randomly if the interface is
changing. etc.