build_objects
A very quick and dirty iterative build setting. It builds object files
and puts them into an objects/ directory with the build.py file. It'll
check the mod times between cpp and o files to decide if it should
build the o again, but that doesn't account for header changes.

I also made the command generation in clang++ a lot more generic,
should make it easier to centeralise and add new compilers.