fixing clean_modules
previously i was relying on cleaning all references to submodules as a
way to ensure they would be imported correctly into .py
configs. it's not easy to control references to a module, so i'm
moving to using importlib.reload.

that requires fixing another design flaw, removing __config_type from
config imports. json & ini configs rely on it, so they've retained it,
i'll potentially remove those in the future too.

* removed del sys.modules
* added importlib.reload
* removed __config_type from configs
* removed __imported_configs